Add other 4 test cases, style the output and refactor names
This commit does a lot of things. It basically makes us ready to present Tuesday. There could probably be other things to add maybe a little more styling for the html or some other stuff but for the most part it could be good. One probalme is that the test cases requirements aren't really done, but I'll do that by the time you read this commit. Or no one is reading this. That could be fine too. Hello if you've made it this far. one thing to note is that the test driver had to be refactored a little with the thing.